Fix: int to long

+++ 0x1A-hash_tables/3-hash_table_set.c
@@ -11,7 +11,7 @@
  */
 int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
-	unsigned int index;
+	unsigned long int index;
 	hash_node_t *array_node = NULL;
 
 	if (ht == NULL || key == NULL)
